Add AEAD, add sb_detached, unit->property tests, more tests, color output (#38)
* Add a utils modules w/ randomVector

It is reasonable to claim this should be `.Random` and produce
`ByteString` instead - the API at this point is arbitrary.

* Add detached signature support.

* Add AEAD

* Fix detached secretbox - check tag size

* Expose AEAD, add tests, revamp testing.

Most the testing was unit tests instead of property tests. This change
makes the test framework generate random tests and includes a number of
checks ot make sure the property should hold.

* Include incorrectly sized tags in test.

* Fix detached API and use unsafe operations safely.

This buildUnsafe* is sharp and pointy. I've increased the number of
default tests - it will cost more time but is entirely worth it.

* Update tests.

* update libsodium version in travis file

* Rename randomVector to randomByteString