Rename byte sizes to what libsodium calls them. Make `Show` safer.
It is probably not a good idea to print all kinds of keys in a `Show`
instance. Now they are hashed to make sure nothing too bad happens,
if they end up in log files etc.

But if you prefer, the shotgun to shoot yourself in the foot with
is provided in `Crypto.Saltine.Core.Utils` as `bin2hex . encode`.